requirements.txt
streamlit
langchain-community
python-dotenv
openai
